+++ lib/templates.ts
@@ -22,6 +22,7 @@ import { isValidISODate } from "./utils/date/isValidISODate";
 import { validateTemplate } from "@lib/utils/form-builder/validate";
 import { dateHasPast } from "@lib/utils";
 import { validateTemplateSize } from "@lib/utils/validateTemplateSize";
+import { NotificationsInterval } from "@gcforms/types";
 
 // ******************************************
 // Internal Module Functions
@@ -46,6 +47,7 @@ const _parseTemplate = (template: {
   closingDate?: Date | null;
   closedDetails?: Prisma.JsonValue | null;
   saveAndResume: boolean;
+  notificationsInterval?: number | null;
 }): FormRecord => {
   return {
     id: template.id,
@@ -79,6 +81,7 @@ const _parseTemplate = (template: {
     }),
     closedDetails: template.closedDetails as ClosedDetails,
     saveAndResume: template.saveAndResume,
+    notificationsInterval: template.notificationsInterval as NotificationsInterval,
   };
 };
 
@@ -96,6 +99,7 @@ export type CreateTemplateCommand = {
   publishReason?: string;
   publishFormType?: string;
   publishDesc?: string;
+  notificationsInterval?: NotificationsInterval;
 };
 
 export type UpdateTemplateCommand = {
@@ -108,6 +112,7 @@ export type UpdateTemplateCommand = {
   publishReason?: string;
   publishFormType?: string;
   publishDesc?: string;
+  notificationsInterval?: NotificationsInterval;
 };
 
 export class InvalidFormConfigError extends Error {
@@ -187,6 +192,9 @@ export async function createTemplate(command: CreateTemplateCommand): Promise<Fo
           connect: { id: command.userID },
         },
         ...(command.formPurpose && { formPurpose: command.formPurpose }),
+        ...(command.notificationsInterval !== undefined && {
+          notificationsInterval: command.notificationsInterval,
+        }),
       },
       select: {
         id: true,
@@ -202,6 +210,7 @@ export async function createTemplate(command: CreateTemplateCommand): Promise<Fo
         publishFormType: true,
         publishDesc: true,
         saveAndResume: true,
+        notificationsInterval: true,
       },
     })
     .catch((e) => prismaErrors(e, null));
@@ -248,6 +257,7 @@ export async function getAllTemplates(options?: {
           publishFormType: true,
           publishDesc: true,
           saveAndResume: true,
+          notificationsInterval: true,
         },
         ...(sortByDateUpdated && {
           orderBy: {
@@ -308,6 +318,7 @@ export async function getAllTemplatesForUser(
           publishFormType: true,
           publishDesc: true,
           saveAndResume: true,
+          notificationsInterval: true,
         },
         ...(sortByDateUpdated && {
           orderBy: {
@@ -371,6 +382,7 @@ export async function getPublicTemplateByID(formID: string): Promise<PublicFormR
           closedDetails: true,
           saveAndResume: true,
           ttl: true,
+          notificationsInterval: true,
         },
       })
       .catch((e) => prismaErrors(e, null));
@@ -540,6 +552,9 @@ export async function updateTemplate(command: UpdateTemplateCommand): Promise<Fo
           securityAttribute: command.securityAttribute as string,
         }),
         ...(command.formPurpose && { formPurpose: command.formPurpose }),
+        ...(command.notificationsInterval !== undefined && {
+          notificationsInterval: command.notificationsInterval as NotificationsInterval,
+        }),
       },
       include: {
         deliveryOption: true,
@@ -956,6 +971,7 @@ export async function updateAssignedUsersForTemplate(
         publishDesc: true,
         users: true,
         saveAndResume: true,
+        notificationsInterval: true,
       },
     })
     .catch((e) => prismaErrors(e, null));
@@ -1050,6 +1066,7 @@ export async function updateFormPurpose(
         publishFormType: true,
         publishReason: true,
         saveAndResume: true,
+        notificationsInterval: true,
       },
     })
     .catch((e) => {
@@ -1109,6 +1126,7 @@ export async function updateFormSaveAndResume(
         publishFormType: true,
         publishReason: true,
         saveAndResume: true,
+        notificationsInterval: true,
       },
     })
     .catch((e) => {
@@ -1177,6 +1195,7 @@ export async function updateResponseDeliveryOption(
         publishFormType: true,
         publishDesc: true,
         saveAndResume: true,
+        notificationsInterval: true,
       },
     })
     .catch((e) => {
@@ -1286,6 +1305,7 @@ export async function deleteTemplate(formID: string): Promise<FormRecord | null>
         publishFormType: true,
         publishDesc: true,
         saveAndResume: true,
+        notificationsInterval: true,
       },
     })
     .catch((e) => prismaErrors(e, null));
@@ -1407,6 +1427,7 @@ export const updateSecurityAttribute = async (formID: string, securityAttribute:
         publishFormType: true,
         publishDesc: true,
         saveAndResume: true,
+        notificationsInterval: true,
       },
     })
     .catch((e) => prismaErrors(e, null));