+++ packages/types/CHANGELOG.md
@@ -5,12 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.6] - 2025-05-06
+## [1.0.8] - 2025-05-20
 
 ### Changed
 
 - Add questionId and tags to ElementProperties
 - Add uuid to FormElement
+
+## [1.0.7] - 2025-04-09
+
+Add form element type for custom json
+
+## [1.0.6] - 2025-04-29
+
+### Added
+
+- New type `NotificationsInterval` also added in FormRecord
 
 ## [1.0.5] - 2025-04-09